+++ src/extension_consts.ml
@@ -27,6 +27,11 @@ module Commands = struct
 
   let open_sandbox_documentation = ocaml_prefixed "open-sandbox-documentation"
 
+  let stop_documentation_server = ocaml_prefixed "stop-documentation-server"
+
+  let generate_sandbox_documentation =
+    ocaml_prefixed "generate-sandbox-documentation"
+
   let open_current_dune_file = ocaml_prefixed "current-dune-file"
 
   let evaluate_selection = ocaml_prefixed "evaluate-selection"